+++ views/layouts/iframe.php
@@ -1,7 +1,7 @@
 <?php
 
 use humhub\assets\AppAsset;
-use humhub\modules\ui\view\components\View;
+use humhub\components\View;
 use yii\helpers\Html;
 
 /* @var $this View */