+++ views/manage/spaceSettings.php
@@ -6,12 +6,10 @@
  * @author [Marc FARRE](https://marc.fun)
  */
 
-use humhub\libs\Html;
+use humhub\components\View;
 use humhub\modules\externalWebsites\models\forms\SpaceSettingsForm;
-use humhub\modules\ui\view\components\View;
-use humhub\widgets\ModalButton;
-use humhub\widgets\ModalDialog;
-use yii\bootstrap\ActiveForm;
+use humhub\widgets\modal\Modal;
+use humhub\widgets\modal\ModalButton;
 
 
 /**
@@ -20,25 +18,12 @@ use yii\bootstrap\ActiveForm;
  */
 ?>
 
-<?php ModalDialog::begin([
-    'header' => Yii::t('ExternalWebsitesModule.base', 'Settings'),
+<?php $form = Modal::beginFormDialog([
+    'title' => Yii::t('ExternalWebsitesModule.base', 'Settings'),
+    'footer' => ModalButton::cancel() . ' ' . ModalButton::save()->submit(),
 ]) ?>
-<div class="modal-body">
-
-    <?php $form = ActiveForm::begin(); ?>
 
     <?= $form->field($model, 'urlToRedirect')->textInput(['placeholder' => 'https://www.my-external-website.tdl', 'autofocus' => '']) ?>
     <?= $form->field($model, 'preventLeavingSpace')->checkbox() ?>
 
-    <br>
-    <div class="input-group-btn">
-        <?= Html::saveButton() ?>
-    </div>
-
-    <?php ActiveForm::end(); ?>
-
-</div>
-<div class="modal-footer">
-    <?= ModalButton::cancel(Yii::t('base', 'Close')) ?>
-</div>
-<?php ModalDialog::end() ?>
+<?php Modal::endFormDialog() ?>